Remove functionality from Importing Controller
Fetching and parsing are moved to other classes.
