+++ src/main/java/org/openrefine/extensions/commons/importer/FileRecord.java
@@ -1,0 +1,17 @@
+package org.openrefine.extensions.commons.importer;
+
+import java.util.List;
+
+//import com.fasterxml.jackson.databind.JsonNode;
+
+/**
+ * This class represents the contents of a record in a project
+ *
+ */
+public class FileRecord {
+
+    String fileName;
+    String mId;
+    List<String> relatedCategories;
+
+}
